+++ src/types/config.ts
@@ -6,7 +6,7 @@ import SendmailTransport from 'nodemailer/lib/sendmail-transport';
 import SESTransport from 'nodemailer/lib/ses-transport';
 import SMTPTransport from 'nodemailer/lib/smtp-transport';
 import StreamTransport from 'nodemailer/lib/stream-transport';
-import { ConsentConfig } from './typings';
+import { ConsentConfig, PooledSMTPOptions } from './typings';
 
 export interface TestConfig {
   /** Use a stub transport so no email is actually sent. Default: false */
@@ -170,24 +170,67 @@ export interface DBServerConfig {
   designDocDir?: string;
 }
 
+export interface EmailTemplateConfig {
+  /**
+   * Customize the templates for the emails that `couch-auth` sends out.
+   * Otherwise, the defaults located in `./templates/email` will be used.
+   * The following templates are used by `couch-auth`:
+   * - `'confirmEmail'`
+   * - `'confirmEmailChange'`
+   * - `'forgotPassword'`
+   * - `'modifiedPassword'`
+   * - `'signupExistingEmail'`
+   * - `'forgotUsername'`
+   *
+   * If a template named `base.njk` exists in the template folder, it will be
+   * used to send out both a HTML and a plain text version based on the contents
+   * of `${template}.njk` templates.
+   * But if both (or any) of `${template}.html.njk` and `${template}.text.njk`
+   * exists, they will be used instead.
+   *
+   * Basic markdown styling is supported:
+   * - `[]()` for URLs
+   * - `_` or `*` for italic
+   * - `**` for bold
+   *
+   * You can add additional templates and send them out via `sendEmail()`, using
+   * the same templating logic.
+   */
+  templates?: Record<string, EmailTemplate>;
+  /**
+   * Folder path relative to which the email templates are located.
+   * If not specified, `./templates/email` is used
+   */
+  folder?: string;
+  /**
+   * Anything defined here will be available as `data. ...` within `nunjucks`.
+   * If the same property is defined both here _and_ in the `EmailTemplate`,
+   * the latter takes precedence.
+   */
+  data?: Record<string, any>;
+}
+
 /**
- * Configure templates that are sent out by superlogin automatically or
- * on-demand when using `superlogin.sendEmail`.
+ * Configure templates that are sent out by `couch-auth` automatically or
+ * on-demand when using ``couch-auth`.sendEmail`.
  */
 export interface EmailTemplate {
   /** The subject for the sent out email */
   subject: string;
-  /** The formats in which the email should be send out */
-  formats?: Array<'text' | 'html'>;
   /**
-   * The paths (relative to where your config file) where the templates for the
-   * different formats are is located (in the same order)
+   * Additional data that can be accessed with `data.` in the nunjucks template.
+   * Note that `paragraphs` is reserved internally when using the hierarchical
+   * template logic!
    */
-  templates?: string[];
-  /** @deprecated The format of the email */
-  format?: 'text' | 'html';
-  /** @deprecated The path of the template */
-  template?: string;
+  data?: Record<string, any>;
+}
+
+export interface RetryMailOptions {
+  /** Retry at most this many times */
+  maxRetries: number;
+  /** Initial amount of seconds to wait. Next attempt will be made after 2x the
+   * previous waiting time. */
+  initialBackoffSeconds: number;
 }
 
 /** Configure how [nodemailer](https://nodemailer.com/about/) sends mails. */
@@ -205,19 +248,22 @@ export interface MailerConfig {
     | SESTransport
     | Transport;
   /**
-   * If you do not use a custom `transport`, these are your SMTP credentials.
+   * If you do not use a custom `transport`, these are your SMTP credentials and
+   * additional options passed to `createTransport`.
    *
    * See https://nodemailer.com/smtp/#examples for details.
    */
-  options?: SMTPTransport.Options;
+  options?: SMTPTransport.Options & PooledSMTPOptions;
   /**
    * Additional message fields, e.g. `replyTo` and `cc`.
    * Note that `to`, `from`, `subject`, `html` and `text` are expected to be
-   * handled by `superlogin` instead.
+   * handled by ``couch-auth`` instead.
    *
    * See https://nodemailer.com/message/ for details.
    */
   messageConfig?: Mail.Options;
+  /** If the call to nodemailer rejected with an error, retry the mail delivery */
+  retryOnError?: RetryMailOptions;
 }
 
 export interface DefaultDBConfig {
@@ -307,10 +353,12 @@ export interface ProviderConfig {
    */
   stateRequired: boolean;
   /**
-   * Custom template for the redirect callback, this needs to pass the data received by the provider authentication
-   * back to the parent window, you can copy the default from `templates/oauth/auth-callback` but modify the
-   * second parameter of the function postMessage, that is the targetOrigin, with the origin of your page server
-   * to avoid posting the data to any potencial malicious site.
+   * Custom `nunjucks` template for the redirect callback, this needs to pass
+   * the data received by the provider authentication back to the parent window,
+   * you can copy the default from `templates/oauth/authCallback` but modify
+   * the second parameter of the function postMessage, that is the targetOrigin,
+   * with the origin of your page server to avoid posting the data to any
+   * potencial malicious site.
    *
    * In the template you have access to
    *  - error: message in case anything went wrong.
@@ -332,20 +380,8 @@ export interface Config {
   dbServer: DBServerConfig;
   /** Configure how mails are sent out to users */
   mailer?: MailerConfig;
-  /**
-   * Customize the templates for the emails that SuperLogin sends out.
-   * Otherwise, the defaults located in `./templates/email` will be used.
-   * The following templates are used by Superlogin:
-   * - `'confirmEmail'`
-   * - `'confirmEmailChange'`
-   * - `'forgotPassword'`
-   * - `'modifiedPassword'`
-   * - `'signupExistingEmail'`
-   * - `'forgotUsername'`
-   *
-   * You can add additional templates and send them out via `sendEmail()`.
-   */
-  emails?: Record<string, EmailTemplate>;
+  /** Configure the email templates, the folder location + additional data */
+  emailTemplates?: EmailTemplateConfig;
   /** Custom settings to manage personal databases for your users */
   userDBs?: UserDBConfig;
   /** OAuth 2 providers */